fix check subscribe channel

+++ handlers/user/channel.py
@@ -14,12 +14,16 @@ async def check_subscribe(call_data: types.CallbackQuery) -> None:
     user_id = call_data.from_user.id
     user_language = await get_language(user_id)
     channel_id = int(call_data.data.replace('check_subscribe ', ' '))
-    print(channel_id)
+
     try:
         member = await call_data.bot.get_chat_member(channel_id, user_id)
-        print(member)
-        await call_data.message.delete()
-        await subscribe_user_to_channel(user_id)
+        if member.status == 'member':
+            await call_data.message.delete()
+            await subscribe_user_to_channel(user_id)
+        else:
+            text_answer = config.messages[user_language]['channel']['not_found']
+            await call_data.answer(text_answer, show_alert=True)
+            return
     except BadRequest:  # если участник не найден
         text_answer = config.messages[user_language]['channel']['not_found']
         await call_data.answer(text_answer, show_alert=True)
